+++ plex-like-front/src/components/Series/Series.tsx
@@ -1,93 +1,42 @@
-import { alpha } from "@mui/material/styles";
-import { Box, Button, SxProps, Typography } from "@mui/material";
-import theme from "../../assets/Themes/Theme";
-import moment from "moment";
-import PlayCircleOutlineIcon from "@mui/icons-material/PlayCircleOutline";
-import Episode from "./components/episode/Episode";
-import { saison } from "./components/episode/components/SaisonMenu";
-import AdditionalInfo from "./components/AdditionalInfo";
-import Presentation from "./components/Presentation";
-import ExpandMoreIcon from "@mui/icons-material/ExpandMore";
 import { useState } from "react";
-import { styleSeriesHover } from "./components/StyleSerieHover";
-import { StyleSerieMore } from "./components/StyleSerieMore";
+import SerieHover from "./components/SerieHover/SerieHover";
+import SerieMore from "./components/SerieMore/SerieMore";
+import SerieInformation from "./components/SerieInfo/SerieInformation";
+import { ISerieInfo } from "../../type/ISerieInfo";
+import { ISaison } from "../../type/ISaison";
 
 interface ISeries {
-  title: string;
-  poster: string;
-  description: string;
-  date: string;
-  realisator: string;
-  actor: string[];
-  publicmark: number;
-  genre: string;
-  saisons: saison[];
+  serieInfo: ISerieInfo;
+  saisons: ISaison[];
 }
 const Series = ({
-  title,
-  poster,
-  description,
-  date,
-  realisator,
-  actor,
-  publicmark,
-  genre,
+  serieInfo,
   saisons,
 }: ISeries): JSX.Element => {
   const [popup, setPopup] = useState(false);
  
-  let styleSeries: SxProps;
-  styleSeries= popup? StyleSerieMore :styleSeriesHover
-
+  if (popup){
   return (
-      <Box className="Series" sx={styleSeries}>
-        <Box
-          className="poster"
-          component="img"
-          sx={{ maxWidth: "10rem" }}
-          alt="logo"
-          src={poster}
-        />
-        <Typography className="titleDate">
-          {title} - {moment(date).format("YYYY")}
-        </Typography>
-        <Box className="presentation">
-          <Presentation
-            title={title}
-            description={description}
-            date={date}
-            genre={genre}
-          />
-        </Box>
-        <Box className="additionalInfo">
-          <AdditionalInfo
-            date={date}
-            realisator={realisator}
-            actor={actor}
-            publicmark={publicmark}
-          />
-        </Box>
-        <Button
-          className="playButton"
-          sx={{ width: "100%", height: "100%" }}
-          variant="text"
-        >
-          <PlayCircleOutlineIcon sx={{ transform: "scale(2.5)" }} />
-        </Button>
-        <Button
-          className="moreButton"
-          onClick={() => {
-            setPopup(true);
-          }}
-          sx={{ minHeight: "1em", minWidth: "1em" }}
-          variant="text"
-        >
-          <ExpandMoreIcon sx={{ transform: "scale(2.5)" }} />
-        </Button>
-        {popup && <Box className="episode" visibility="hidden" >
-          <Episode saisons={saisons} numSaison={1}/>
-        </Box>}
-      </Box>
+    <>
+    <SerieMore 
+      serieInfo={serieInfo}
+      saisons={saisons}
+      setPopup={setPopup}/>
+    <SerieHover 
+      serieInfo={serieInfo}
+      saisons={saisons}
+      setPopup={setPopup}/>
+    </>
   );
+}
+else{
+  return (
+    <SerieHover  
+    serieInfo={serieInfo}
+    saisons={saisons}
+    setPopup={setPopup}/>
+
+  );
+}
 };
 export default Series;
