+++ plex-like-front/src/type/ISaison.tsx
@@ -1,0 +1,7 @@
+
+export interface ISaison {
+    numSaison: number;
+    nameEpisode: string[];
+    durationEpisode: number[];
+    descriptionEpisode: string[];
+};
