+++ plex-like-front/src/type/ISerieInfo.tsx
@@ -1,0 +1,10 @@
+export interface ISerieInfo {
+    title: string;
+    poster: string;
+    description: string;
+    date: string;
+    realisator: string;
+    actor: string[];
+    publicmark: number;
+    genre: string;
+  };
